Extend viewroutes to state components.
Big component refactor to make file/compnoent names and locations more sensible.

Move all sorting behavior into api, components never reformat data unless user does so.

Repeat initial hook pattern on state side (useStateState).
