+++ src/components/federal/CongressSelector.tsx
@@ -5,23 +5,22 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-
-import { getChamberDisplayName, getCongressDisplayName } from "@/lib/display";
+import { getValidCongressDisplayName } from "@/lib/display";
 import type { ValidCongress } from "@/services/api.types";
 
 interface CongressSelectorProps {
-  validCongresses: ValidCongress[];
+  validCongresses: Array<ValidCongress>;
   selectedCongress: number;
   onCongressChange: (congress: number) => void;
+  isLoading?: boolean;
 }
 
 export function CongressSelector({
   validCongresses,
   selectedCongress,
   onCongressChange,
+  isLoading = false,
 }: CongressSelectorProps) {
-  // Sort congresses in descending order and find the current congress's chamber
-  const sortedCongresses = [...validCongresses].sort(([a], [b]) => b - a);
   const currentChamber =
     validCongresses.find(([congress]) => congress === selectedCongress)?.[1] ||
     "H";
@@ -30,17 +29,17 @@ export function CongressSelector({
     <Select
       value={selectedCongress.toString()}
       onValueChange={(value) => onCongressChange(parseInt(value))}
+      disabled={isLoading}
     >
       <SelectTrigger className="w-full">
         <SelectValue>
-          {getCongressDisplayName(selectedCongress)}{" "}
-          {getChamberDisplayName(currentChamber)}
+          {getValidCongressDisplayName([selectedCongress, currentChamber])}
         </SelectValue>
       </SelectTrigger>
       <SelectContent>
-        {sortedCongresses.map(([congress, chamber]) => (
+        {validCongresses.map(([congress, chamber]) => (
           <SelectItem key={congress} value={congress.toString()}>
-            {getCongressDisplayName(congress)} {getChamberDisplayName(chamber)}
+            {getValidCongressDisplayName([congress, chamber])}
           </SelectItem>
         ))}
       </SelectContent>
