--- src/components/state/StateRoot.tsx
@@ -1,71 +1,0 @@
-import { StateControlPanel } from "@/components/state/StateControlPanel";
-import { StateScorecardView } from "@/components/state/StateScorecardView";
-import { StateTableView } from "@/components/state/StateTableView";
-import { useStateState } from "@/hooks/useStateState";
-import type { ViewLevel } from "@/lib/types";
-import { LevelToggle } from "@/components/shared/LevelToggle";
-import { StateTableGlossary } from "@/components/state/StateTableGlossary";
-
-type StateRootProps = {
-  level: ViewLevel;
-  setLevel: (level: ViewLevel) => void;
-};
-function StateRoot({ level, setLevel }: StateRootProps) {
-  const stateState = useStateState();
-
-  if (stateState.selectedStateLegislator) {
-    return (
-      <div className="min-h-screen bg-background">
-        <div className="container mx-auto px-4 py-8 max-w-[1400px]">
-          <StateScorecardView
-            legislator={stateState.selectedStateLegislator}
-            scorecard={stateState.selectedStateScorecard}
-            initialTerm={stateState.selectedStateTerm}
-            onBack={() => {
-              stateState.setSelectedStateLegislator(null);
-              stateState.setSelectedStateScorecard(null);
-            }}
-          />
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="min-h-screen bg-background">
-      <div className="relative flex flex-col items-center px-4 py-8">
-        <div className="w-full max-w-[1200px]">
-          <div className="grid gap-x-6 gap-y-4 grid-cols-1 grid-rows-[min-content_min-content_min-content] grid-areas-mobile md:grid-cols-[1fr_2fr] md:grid-rows-[min-content_1fr] md:grid-areas-desktop">
-            <div className="grid-in-a space-y-4">
-              <LevelToggle level={level} onLevelChange={setLevel} />
-              <StateControlPanel
-                selectedState={stateState.selectedState}
-                selectedTerm={stateState.selectedTerm}
-                selectedChamber={stateState.stateChamber}
-                onStateSelect={stateState.setSelectedState}
-                onTermChange={stateState.setSelectedTerm}
-                onChamberChange={stateState.setStateChamber}
-                onSearchChange={stateState.setStateSearchTerm}
-              />
-            </div>
-            <div className="grid-in-b space-y-4 min-h-[600px]">
-              <h1 className="text-3xl font-bold">DYNAMIC TITLE</h1>
-              <StateTableView
-                selectedState={stateState.selectedState}
-                selectedTerm={stateState.selectedTerm}
-                chamber={stateState.stateChamber}
-                searchTerm={stateState.stateSearchTerm}
-                onLegislatorSelect={stateState.handleStateLegislatorSelect}
-              />
-            </div>
-            <div className="grid-in-c space-y-4">
-              <StateTableGlossary />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export default StateRoot;
