+++ src/lib/display.ts
@@ -1,4 +1,5 @@
 import type { FederalChamberInitial, StateChamber } from "@/lib/types";
+import { ValidCongress, ValidTerm } from "@/services/api.types";
 
 export const getCongressDisplayName = (congress: number): string => {
   const j = congress % 10;
@@ -9,6 +10,15 @@ export const getCongressDisplayName = (congress: number): string => {
   else if (j === 3 && k !== 13) suffix = "rd";
   return `${congress}${suffix}`;
 };
+
+export const getValidCongressDisplayName = ([
+  congress,
+  chamber,
+]: ValidCongress): string => {
+  const chamberName = getChamberDisplayName(chamber);
+  return `${getCongressDisplayName(congress)} ${chamberName}`;
+};
+
 export const getStateLocationText = (district: number | null): string =>
   `District ${district}`;
 export const getFederalLocationText = (
@@ -55,18 +65,12 @@ export const getStateChamberDisplayName = (chamber: StateChamber): string => {
   return chamber === "upper" ? "Upper" : "Lower";
 };
 
-export interface TermDisplayInfo {
-  startYear: number;
-  endYear: number;
-  chamber: StateChamber;
-}
-
-export const getTermDisplayName = (term: TermDisplayInfo): string => {
+export const getValidTermDisplayName = (term: ValidTerm): string => {
   const chamberName = getStateChamberDisplayName(term.chamber);
   return `${term.startYear}-${term.endYear} ${chamberName}`;
 };
 
-export const getTermValue = (term: {
+export const getTermDisplayName = (term: {
   startYear: number;
   endYear: number;
 }): string => {
